insect-3d-tracking/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
python-socketio>=5.8.0
websockets>=11.0.0
pytest-mock>=3.6.0
python-multipart>=0.0.6
httpx>=0.23.0
aiofiles>=23.1.0
